+++ battle_chess/game_manager.py
